+++ wgpu-core/src/command/compute_command.rs
@@ -1,3 +1,5 @@
+use alloc::vec::Vec;
+
 #[cfg(feature = "serde")]
 use crate::command::serde_object_reference_struct;
 use crate::command::{ArcReferences, ReferenceType};
@@ -64,6 +66,11 @@ pub enum ComputeCommand<R: ReferenceType> {
     },
 
     EndPipelineStatisticsQuery,
+
+    TransitionResources {
+        buffer_transitions: Vec<wgt::BufferTransition<R::Buffer>>,
+        texture_transitions: Vec<wgt::TextureTransition<R::TextureView>>,
+    },
 }
 
 /// cbindgen:ignore